+++ lib/assets/word_data.dart
@@ -1,0 +1,39 @@
+import 'package:flash_card_app/model/category.dart';
+
+const wordData = [
+  {
+    "category": CategoryCode.color,
+    "korWord": "빨간색",
+    "engWord": "Red",
+    "image": "https://example.com/images/hello.jpg",
+    "history": []
+  },
+  {
+    "category": CategoryCode.animal,
+    "korWord": "고양이",
+    "engWord": "Cat",
+    "image": "https://example.com/images/cat.jpg",
+    "history": []
+  },
+  {
+    "category": CategoryCode.animal,
+    "korWord": "강아지",
+    "engWord": "Dog",
+    "image": "https://example.com/images/love.jpg",
+    "history": []
+  },
+  {
+    "category": CategoryCode.color,
+    "korWord": "파란색",
+    "engWord": "Blue",
+    "image": "https://example.com/images/music.jpg",
+    "history": []
+  },
+  {
+    "category": CategoryCode.plant,
+    "korWord": "나무",
+    "engWord": "Tree",
+    "image": "https://example.com/images/happiness.jpg",
+    "history": []
+  }
+];
